+++ lib/widgets/common/app_bar/tag_button.dart
@@ -1,34 +1,31 @@
 import 'package:aves/model/actions/entry_actions.dart';
-import 'package:aves/model/filters/album.dart';
 import 'package:aves/model/filters/filters.dart';
+import 'package:aves/model/filters/tag.dart';
 import 'package:aves/model/settings/settings.dart';
 import 'package:aves/model/source/collection_source.dart';
-import 'package:aves/widgets/common/app_bar/quick_choosers/album_chooser.dart';
 import 'package:aves/widgets/common/app_bar/quick_choosers/chooser_button.dart';
 import 'package:aves/widgets/common/app_bar/quick_choosers/filter_chooser.dart';
+import 'package:aves/widgets/common/app_bar/quick_choosers/tag_chooser.dart';
 import 'package:aves/widgets/common/providers/media_query_data_provider.dart';
 import 'package:aves/widgets/filter_grids/common/filter_nav_page.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-class MoveButton extends ChooserQuickButton<String> {
-  final bool copy;
-
-  const MoveButton({
+class TagButton extends ChooserQuickButton<CollectionFilter> {
+  const TagButton({
     super.key,
-    required this.copy,
     super.chooserPosition,
     super.onChooserValue,
     required super.onPressed,
   });
 
   @override
-  State<MoveButton> createState() => _MoveButtonState();
+  State<TagButton> createState() => _TagButtonState();
 }
 
-class _MoveButtonState extends ChooserQuickButtonState<MoveButton, String> {
-  EntryAction get action => widget.copy ? EntryAction.copy : EntryAction.move;
+class _TagButtonState extends ChooserQuickButtonState<TagButton, CollectionFilter> {
+  EntryAction get action => EntryAction.editTags;
 
   @override
   Widget get icon => action.getIcon();
@@ -38,14 +35,14 @@ class _MoveButtonState extends ChooserQuickButtonState<MoveButton, String> {
 
   @override
   Widget buildChooser(Animation<double> animation) {
-    final options = settings.recentDestinationAlbums;
+    final options = settings.recentTags;
     final takeCount = FilterQuickChooser.maxOptionCount - options.length;
     if (takeCount > 0) {
       final source = context.read<CollectionSource>();
-      final filters = source.rawAlbums.whereNot(options.contains).map((album) => AlbumFilter(album, null)).toSet();
+      final filters = source.sortedTags.map(TagFilter.new).whereNot(options.contains).toSet();
       final allMapEntries = filters.map((filter) => FilterGridItem(filter, source.recentEntry(filter))).toList();
       allMapEntries.sort(FilterNavigationPage.compareFiltersByDate);
-      options.addAll(allMapEntries.take(takeCount).map((v) => v.filter.album));
+      options.addAll(allMapEntries.take(takeCount).map((v) => v.filter));
     }
 
     return MediaQueryDataProvider(
@@ -53,7 +50,7 @@ class _MoveButtonState extends ChooserQuickButtonState<MoveButton, String> {
         opacity: animation,
         child: ScaleTransition(
           scale: animation,
-          child: AlbumQuickChooser(
+          child: TagQuickChooser(
             valueNotifier: chooserValueNotifier,
             options: widget.chooserPosition == PopupMenuPosition.over ? options.reversed.toList() : options,
             pointerGlobalPosition: pointerGlobalPosition,